build error fix

+++ exercises/joseph_torsney/ex02_oo_basics/src/square.h
@@ -12,5 +12,5 @@ class Square: public Shape {
     double getPerimeter() const override;
 
   private:
-    const int m_side;
+    const double m_side;
 };
